+++ spec/osm/section_spec.rb
@@ -24,6 +24,7 @@ describe "Section" do
       :myscout_events_expires => (Date.today + 61).strftime('%Y-%m-%d'),
       :myscout_badges_expires => (Date.today + 62).strftime('%Y-%m-%d'),
       :myscout_programme_expires => (Date.today + 63).strftime('%Y-%m-%d'),
+      :myscout_details_expires => (Date.today + 64),
       :myscout_events => true,
       :myscout_badges => true,
       :myscout_programme => true,
@@ -38,6 +39,8 @@ describe "Section" do
       :myscout_event_reminder_frequency => 5,
       :myscout_payment_reminder_count => 6,
       :myscout_payment_reminder_frequency => 7,
+      :myscout_details => true,
+      :myscout_details_email_changes_to => 'notify-changes-to@example.com',
       :sms_sent_test => true,
       :sms_messages_sent => 8,
       :sms_messages_remaining => 9,
@@ -64,6 +67,7 @@ describe "Section" do
     section.myscout_events_expires.should == Date.today + 61
     section.myscout_badges_expires.should == Date.today + 62
     section.myscout_programme_expires.should == Date.today + 63
+    section.myscout_details_expires.should == Date.today + 64
     section.myscout_events.should == true
     section.myscout_badges.should == true
     section.myscout_programme.should == true
@@ -78,6 +82,8 @@ describe "Section" do
     section.myscout_event_reminder_frequency.should == 5
     section.myscout_payment_reminder_count.should == 6
     section.myscout_payment_reminder_frequency.should == 7
+    section.myscout_details.should == true
+    section.myscout_details_email_changes_to.should == 'notify-changes-to@example.com'
     section.sms_sent_test.should == true
     section.sms_messages_sent.should == 8
     section.sms_messages_remaining.should == 9
@@ -98,6 +104,7 @@ describe "Section" do
     section.flexi_records.should == []
     section.myscout_email_address_from.should == ''
     section.myscout_email_address_copy.should == ''
+    section.myscout_details_email_changes_to.should == ''
     section.sms_sent_test.should == false
     section.sms_messages_sent.should == 0
     section.sms_messages_remaining.should == 0
@@ -108,7 +115,7 @@ describe "Section" do
 
     before :each do
       body = [
-        {"sectionConfig"=>"{\"subscription_level\":1,\"subscription_expires\":\"2013-01-05\",\"sectionType\":\"beavers\",\"columnNames\":{\"column_names\":\"names\"},\"numscouts\":10,\"hasUsedBadgeRecords\":true,\"hasProgramme\":true,\"extraRecords\":[{\"name\":\"Flexi Record 1\",\"extraid\":\"111\"}],\"wizard\":\"false\",\"fields\":{\"fields\":true},\"intouch\":{\"intouch_fields\":true},\"mobFields\":{\"mobile_fields\":true},\"gocardless\":\"true\",\"portal\":{\"paymentRemindFrequency\":\"7\",\"paymentRemindCount\":\"6\",\"eventRemindFrequency\":\"5\",\"eventRemindCount\":\"4\",\"badgesPartial\":1,\"programmeTimes\":1,\"programmeSummary\":1,\"emailAddress\":\"send_from@example.com\",\"emailAddressCopy\":null,\"payments\":1,\"badges\":1,\"emails\":{\"email1\":\"true\",\"email2\":\"false\"},\"events\":1,\"programme\":1},\"portalExpires\":{\"events\":\"2013-01-06\",\"eventsA\":1,\"badges\":\"2013-01-07\",\"badgesA\":1,\"programme\":\"2013-01-08\",\"programmeA\":1},\"hasSentTestSMS\":true,\"sms_remaining\":8,\"sms_sent\":9}", "groupname"=>"3rd Somewhere", "groupid"=>"3", "groupNormalised"=>"1", "sectionid"=>"1", "sectionname"=>"Section 1", "section"=>"beavers", "isDefault"=>"1", "permissions"=>{"badge"=>10, "member"=>20, "user"=>100, "register"=>100, "contact"=>100, "programme"=>100, "originator"=>1, "events"=>100, "finance"=>100, "flexi"=>100}},
+        {"sectionConfig"=>"{\"subscription_level\":1,\"subscription_expires\":\"2013-01-05\",\"sectionType\":\"beavers\",\"columnNames\":{\"column_names\":\"names\"},\"numscouts\":10,\"hasUsedBadgeRecords\":true,\"hasProgramme\":true,\"extraRecords\":[{\"name\":\"Flexi Record 1\",\"extraid\":\"111\"}],\"wizard\":\"false\",\"fields\":{\"fields\":true},\"intouch\":{\"intouch_fields\":true},\"mobFields\":{\"mobile_fields\":true},\"gocardless\":\"true\",\"portal\":{\"paymentRemindFrequency\":\"7\",\"paymentRemindCount\":\"6\",\"eventRemindFrequency\":\"5\",\"eventRemindCount\":\"4\",\"badgesPartial\":1,\"programmeTimes\":1,\"programmeSummary\":1,\"details\":1,\"contactNotificationEmail\":\"notify-changes-to@example.com\",\"emailAddress\":\"send_from@example.com\",\"emailAddressCopy\":null,\"payments\":1,\"badges\":1,\"emails\":{\"email1\":\"true\",\"email2\":\"false\"},\"events\":1,\"programme\":1},\"portalExpires\":{\"events\":\"2013-01-06\",\"eventsA\":1,\"badges\":\"2013-01-07\",\"badgesA\":1,\"programme\":\"2013-01-08\",\"programmeA\":1,\"details\":\"2013-01-09\",\"detailsA\":1},\"hasSentTestSMS\":true,\"sms_remaining\":8,\"sms_sent\":9}", "groupname"=>"3rd Somewhere", "groupid"=>"3", "groupNormalised"=>"1", "sectionid"=>"1", "sectionname"=>"Section 1", "section"=>"beavers", "isDefault"=>"1", "permissions"=>{"badge"=>10, "member"=>20, "user"=>100, "register"=>100, "contact"=>100, "programme"=>100, "originator"=>1, "events"=>100, "finance"=>100, "flexi"=>100}},
         {"sectionConfig"=>"{\"subscription_level\":3,\"subscription_expires\":\"2013-01-05\",\"sectionType\":\"cubs\",\"columnNames\":{\"phone1\":\"Home Phone\",\"phone2\":\"Parent 1 Phone\",\"address\":\"Member's Address\",\"phone3\":\"Parent 2 Phone\",\"address2\":\"Address 2\",\"phone4\":\"Alternate Contact Phone\",\"subs\":\"Gender\",\"email1\":\"Parent 1 Email\",\"medical\":\"Medical / Dietary\",\"email2\":\"Parent 2 Email\",\"ethnicity\":\"Gift Aid\",\"email3\":\"Member's Email\",\"religion\":\"Religion\",\"email4\":\"Email 4\",\"school\":\"School\"},\"numscouts\":10,\"hasUsedBadgeRecords\":true,\"hasProgramme\":true,\"extraRecords\":[],\"wizard\":\"false\",\"fields\":{\"email1\":true,\"email2\":true,\"email3\":true,\"email4\":false,\"address\":true,\"address2\":false,\"phone1\":true,\"phone2\":true,\"phone3\":true,\"phone4\":true,\"school\":false,\"religion\":true,\"ethnicity\":true,\"medical\":true,\"patrol\":true,\"subs\":true,\"saved\":true},\"intouch\":{\"address\":true,\"address2\":false,\"email1\":false,\"email2\":false,\"email3\":false,\"email4\":false,\"phone1\":true,\"phone2\":true,\"phone3\":true,\"phone4\":true,\"medical\":false},\"mobFields\":{\"email1\":false,\"email2\":false,\"email3\":false,\"email4\":false,\"address\":true,\"address2\":false,\"phone1\":true,\"phone2\":true,\"phone3\":true,\"phone4\":true,\"school\":false,\"religion\":false,\"ethnicity\":true,\"medical\":true,\"patrol\":true,\"subs\":false},\"hasSentTestSMS\":true,\"sms_remaining\":8,\"sms_sent\":9}", "groupname"=>"1st Somewhere", "groupid"=>"1", "groupNormalised"=>"1", "sectionid"=>"2", "sectionname"=>"Section 2", "section"=>"cubs", "isDefault"=>"0", "permissions"=>{"badge"=>100, "member"=>100, "user"=>100, "register"=>100, "contact"=>100, "programme"=>100, "originator"=>1, "events"=>100, "finance"=>100, "flexi"=>100}}
       ]
       FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/api.php?action=getUserRoles", :body => body.to_json)
@@ -138,6 +145,7 @@ describe "Section" do
         section.myscout_events_expires.should == Date.new(2013, 1, 6)
         section.myscout_badges_expires.should == Date.new(2013, 1, 7)
         section.myscout_programme_expires.should == Date.new(2013, 1, 8)
+        section.myscout_details_expires.should == Date.new(2013, 1, 9)
         section.myscout_events.should == true
         section.myscout_badges.should == true
         section.myscout_programme.should == true
@@ -152,6 +160,8 @@ describe "Section" do
         section.myscout_event_reminder_frequency.should == 5
         section.myscout_payment_reminder_count.should == 6
         section.myscout_payment_reminder_frequency.should == 7
+        section.myscout_details.should == true
+        section.myscout_details_email_changes_to.should == 'notify-changes-to@example.com'
         section.sms_sent_test.should == true
         section.sms_messages_remaining.should == 8
         section.sms_messages_sent.should == 9